requirements.txt
celery==5.0.5
python-decouple==3.4
redis==3.5.3
selenium==3.141.0
